requirements.txt
absl-py==2.1.0
astroid==3.3.6
black==24.10.0
click==8.1.7
dill==0.3.9
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.68.1
isort==5.13.2
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.1
pylint==3.3.2
pytoml==0.1.21
PyYAML==6.0.2
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
Werkzeug==3.1.3
